Add files via upload
This notebook is for the Machine Learning mini-projects which I completed for WorldQuant University Data Science Module Unit II: Machine Learning and Statistical Analysis course as they were the assignments for us the students. 
"The objective of this mini-project is to exercise your ability to create effective machine learning models for making predictions. We will be working with nursing home inspection data from the United States, predicting which providers may be fined and for how much."